backend/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
gunicorn==21.2.0
requests==2.31.0

pdfkit==1.0.0
PyPDF2==3.0.1
selenium==4.10.0
webdriver-manager==4.0.0

google-api-core==2.11.1
google-auth>=2.22.0
google-cloud-secret-manager==2.16.3
google-cloud-storage>=2.14.0
google-cloud-aiplatform>=1.29.0
google-cloud-core==2.3.3
google-cloud-discoveryengine
google-cloud-documentai==2.20.0

langchain==0.0.219
langchain-google-vertexai==1.0.4

